refactor(ios): split AppDelegate.swift into RegexPatterns, OCRService, PhotoScanner
AppDelegate was a 835-line monolith mixing regex patterns, OCR logic,
photo scanning, and channel wiring. Extract into focused modules:

- RegexPatterns.swift: all pre-compiled NSRegularExpression patterns
- OCRService.swift: Vision OCR, text extraction, date/calendar helpers
- PhotoScanner.swift: PHPhotoLibrary scanning, progress callbacks,
  single-image scan, asset load/delete

AppDelegate now contains only channel setup and delegation (~69 lines).
No behavior changes — pure refactor.
